Move `ValueTracker` to separate file
