+++ src/common/value-tracker.ts
@@ -3,27 +3,27 @@ import { EventProducer } from 'fun-events';
 /**
  * Value accessor and changes tracker.
  */
-export abstract class ValueTracker<T = any, O = T> {
+export abstract class ValueTracker<T = any, N extends T = T> {
 
   /**
    * Value changes event producer.
    *
    * The registered event consumers receive new and old values as arguments.
    */
-  abstract readonly on: EventProducer<(this: void, newValue: T, oldValue: O) => void>;
+  abstract readonly on: EventProducer<(this: void, newValue: N, oldValue: T) => void>;
 
   /**
    * Reads the tracked value.
    *
    * @returns The value.
    */
-  abstract get it(): O;
+  abstract get it(): T;
 
   /**
    * Updates the tracked value.
    *
    * @param value New value.
    */
-  abstract set it(value: O);
+  abstract set it(value: T);
 
 }
