+++ src/index.ts
@@ -1,3 +1,4 @@
+export * from './common';
 export * from './tree';
 
 import '@wesib/wesib';
