+++ src/tree/attribute-tracker.ts
@@ -1,12 +1,12 @@
 import { ComponentContext } from '@wesib/wesib';
 import { EventProducer } from 'fun-events';
 import { AttributesObserver } from './attributes-observer';
-import { ValueTracker } from './value-tracker';
+import { ValueTracker } from '../common';
 
 /**
  * @internal
  */
-export class AttributeTracker extends ValueTracker<string, string | null> {
+export class AttributeTracker extends ValueTracker<string | null | string> {
 
   readonly on: EventProducer<(this: void, newValue: string, oldValue: string | null) => void>;
