+++ src/tree/index.ts
@@ -1,3 +1,3 @@
 export * from './component-node';
 export * from './component-tree-support.feature';
-export * from './value-tracker';
+export * from '../common/value-tracker';
